web/phlicker/inner/requirements.txt
flask
flask-login
flask-sqlalchemy

pillow
requests
